+++ frontend/src/components/messages/MessageList.js
@@ -1,13 +1,21 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 
 const MessageList = (props) => {
-  console.log(props);
   return props.messages.map( message => {
     return (
-      <div>
-        <p> {message[props.type].fullname} </p>
-        <p> {message.messages[0].text} </p>
-      </div>
+      <Link className="thread" to={`/${props.type}/messages/${message.id}`}>
+        <div className="thread__picture-container">
+          <img className="thread__user-picture" src={message[props.type].picture}/>
+        </div>
+        <div className="thread__content">
+          <div className="thread__header">
+            <span className="thread__title"> {message[props.type].fullname} </span>
+            <small className="thread__time"> 7:49 PM</small>
+          </div>
+          <p className="thread__body"> {message.messages[0].text} </p>
+        </div>
+      </Link>
     )
   })
 }
